Refactor: 이벤트 핸들러 함수 분리
