+++ .eslintrc.json
@@ -14,6 +14,7 @@
     "import/extensions": "off",
     "no-plusplus": "off",
     "import/no-extraneous-dependencies": ["error", { "devDependencies": true }],
-    "import/prefer-default-export": "off"
+    "import/prefer-default-export": "off",
+    "no-alert": "off"
   }
 }
